http://jira.codehaus.org/browse/MODFORJ-189: Started by removing the checks for null pointers in the translators, so that the validators in the business domain classes get their chance to fire and raise exceptions that give useful information to the client about the errors found. For this to work the parameters for non-nullable properties of setters and constructors do not use primitives anymore.

+++ modules/Examples/RecordShop-ExampleProject/RecordShop-data/src/test/java/org/company/recordshop/AssociationTest.java
@@ -99,7 +99,7 @@ public class AssociationTest extends AbstractDaoTestCase {
         return result;
     }
 
-    protected Product createProduct(String prodNr, int price) {
+    protected Product createProduct(String prodNr, float price) {
         Product result = new Product(prodNr, price, false);
         allProducts.add(result);
         productDao.add(result);
